Add square root NextHop difference formula

+++ src/nexthop-difference-calculator.hpp
@@ -11,7 +11,7 @@
 class NextHopDifferenceCalculator
 {
 public:
-  static int
+  static double
   getDifference(const NextHopSet& lhs, const NextHopSet& rhs);
 };
